backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
pdf2image==1.16.3
PyPDF2==3.0.1
PyMuPDF==1.23.8
reportlab==4.0.7
Pillow==10.1.0
pydantic==2.5.0
sqlalchemy==2.0.23
aiosqlite==0.19.0

